+++ app/components/cards/player-card.tsx
